refactor: Exception 클래스 계층 구조 변경
