+++ packy-common/src/main/java/com/dilly/exception/EntityNotFoundException.java
@@ -1,7 +1,4 @@
-package com.dilly.exception.entitynotfound;
-
-import com.dilly.exception.BusinessException;
-import com.dilly.exception.ErrorCode;
+package com.dilly.exception;
 
 public class EntityNotFoundException extends BusinessException {
