--- packy-common/src/main/java/com/dilly/exception/entitynotfound/MemberNotFoundException.java
@@ -1,14 +1,0 @@
-package com.dilly.exception.entitynotfound;
-
-import com.dilly.exception.ErrorCode;
-
-public class MemberNotFoundException extends EntityNotFoundException {
-
-	public MemberNotFoundException() {
-		super(ErrorCode.MEMBER_NOT_FOUND);
-	}
-
-	public MemberNotFoundException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
