+++ packy-domain/src/main/java/com/dilly/member/adaptor/MemberReader.java
@@ -1,6 +1,7 @@
 package com.dilly.member.adaptor;
 
-import com.dilly.exception.entitynotfound.MemberNotFoundException;
+import com.dilly.exception.EntityNotFoundException;
+import com.dilly.exception.ErrorCode;
 import com.dilly.member.MemberRepository;
 import com.dilly.member.domain.Member;
 import com.dilly.member.domain.Status;
@@ -14,7 +15,9 @@ public class MemberReader {
 	private final MemberRepository memberRepository;
 
 	public Member findById(Long id) {
-		return memberRepository.findById(id).orElseThrow(MemberNotFoundException::new);
+		return memberRepository.findById(id).orElseThrow(
+			() -> new EntityNotFoundException(ErrorCode.MEMBER_NOT_FOUND)
+		);
 	}
 
     public Long countByStatus(Status status) {
